Revert "API Demo: set D-Pad to player number"
On second thought, this requires the send functionality to work *anyways*... It would be better to toggle on an LED, but not all boards have built-in lEDs and even then you need to show 5 states (off/nc, players 1-4) which means 3 LEDs. I'm going to leave this with the 'send' only practical test for now until I can come up with something better. This reverts commit 2e01538960644d5962f57b873901b622aa0fe1c9.