Add joystick functions for separate x and y axes
If for whatever reason you only want to set one axis, you can now do so without having to figure out where the middle of the pre-scaled range is. For now the X/Y range is still being treated the same because this is an edge case, not the main use (i.e. most users will still use joysticks as joysticks).