Flush receive buffer on construction (#50)
On certain platforms this was causing a problem where the library was always one or two packets behind because USB packets were received and buffered by the hardware controller before the library object was constructed and could set its callback. Flushing the receive buffer when the object is constructed appears to solve this issue.